[Update] Bullet Spawner와 Controller 분리
