+++ Assets/ExcelDatas/~$SkillEnchantData.xlsx.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e19f310565c333647bba7d0d56977132
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
