+++ Assets/Scripts/Bullets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fea1709ac6fd8a44dafa86758c04c2fd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
